// src/app/Shared/models/BillingAddress.ts
export default class BillingAddress {
    BillingAddressId: number;
    Name: string;
    CustomerId: number;
    Address1: string;
    Address2: string;
    LandMark: string;
    ContactNo: string;
    Pincode: string;
    StateId: number;
    CountryId: number;
    Latitude: string;
    Longitude: string;
    Active: boolean;
    AddressType: string;
    IsDefault: boolean;
    CreatedDate: Date;
    UpdatedDate: Date;
    StatusCode: number;
    Status: string;
}